new price with GST Added

+++ assets/custom.js
@@ -109,9 +109,9 @@ for (let i = 0, linksLength = links.length ; i < linksLength ; i++) {
 // Tax calculetor for product price
 function priceWIthGST(gst,price){
   var pr = price.split('$')
-  var newPrice = parseFloat(pr[1]);
+  var newPrice = parseInt(pr[1]);
   let wIthGST = newPrice * gst / 100;
-  let newPriceWithGST = parseFloat(pr[1]) + wIthGST.toFixed(2)
+  let newPriceWithGST = parseInt(pr[1]) + wIthGST.toFixed(2)
   return (`$`+ newPriceWithGST)
 }
